+++ backend/src/main/java/com/simulation/trading/crypto/repository/UserRepository.java
@@ -6,6 +6,7 @@ import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 @Repository
@@ -24,15 +25,14 @@ public class UserRepository {
             return Optional.of(jdbcTemplate.queryForObject(sql,
                     (rs, rowNum) -> new User(rs.getInt("id"),
                                                         rs.getString("name"),
-                                                        rs.getDouble("balance")),
+                                                        rs.getBigDecimal("balance")),
                     userID));
         } catch (EmptyResultDataAccessException e){
             return Optional.empty();
         }
     }
 
-    //TODO: return int to indicate num of affected rows, maybe impl check to veify ==1
-    public int updateUserBalance(int userID, double balance) {
+    public int updateUserBalance(int userID, BigDecimal balance) {
         String sql = "UPDATE users SET balance = ? WHERE id = ?";
         return jdbcTemplate.update(sql, balance, userID);
     }
